### 2023-01-26 #5102 @editorialbot
[PRE REVIEW]: mutyper: assigning and summarizing mutation types for analyzing germline mutation spectra
**Submitting author:** @WSDeWitt (<a href="http://orcid.org/0000-0002-6802-9139">William DeWitt</a>)
**Repository:** https://github.com/harrispopgen/mutyper
**Branch with paper.md** (empty if default branch):
**Version:** v0.7.0
