### 2023-02-08 08c1eaa
add missing DOIs to bib, and clean/simplify bib
