+++ paper/paper.bib
@@ -6,9 +6,9 @@
   volume  =  62,
   pages   = "50--57",
   month   =  jun,
-  year    =  2020
+  year    =  2020,
+  doi     = {10.1016/j.gde.2020.05.024}
 }
-
 
 @article{10.1093/gbe/evab104,
     author = {Goldberg, Michael E and Harris, Kelley},
@@ -18,15 +18,9 @@
     number = {1},
     year = {2021},
     month = {05},
-    abstract = "{Great ape clades exhibit variation in the relative mutation rates of different three-base-pair genomic motifs, with closely related species having more similar mutation spectra than distantly related species. This pattern cannot be explained by classical demographic or selective forces, but imply that DNA replication fidelity has been perturbed in different ways on each branch of the great ape phylogeny. Here, we use whole-genome variation from 88 great apes to investigate whether these species’ mutation spectra are broadly differentiated across the entire genome, or whether mutation spectrum differences are driven by DNA compartments that have particular functional features or chromatin states. We perform principal component analysis (PCA) and mutational signature deconvolution on mutation spectra ascertained from compartments defined by features including replication timing and ancient repeat content, finding evidence for consistent species-specific mutational signatures that do not depend on which functional compartments the spectra are ascertained from. At the same time, we find that many compartments have their own characteristic mutational signatures that appear stable across the great ape phylogeny. For example, in a mutation spectrum PCA compartmentalized by replication timing, the second principal component explaining 21.2\\% of variation separates all species’ late-replicating regions from their early-replicating regions. Our results suggest that great ape mutation spectrum evolution is not driven by epigenetic changes that modify mutation rates in specific genomic regions, but instead by trans-acting mutational modifiers that affect mutagenesis across the whole genome fairly uniformly.}",
     issn = {1759-6653},
     doi = {10.1093/gbe/evab104},
-    url = {https://doi.org/10.1093/gbe/evab104},
-    note = {evab104},
-    eprint = {https://academic.oup.com/gbe/article-pdf/14/1/evab104/42081122/evab104.pdf},
 }
-
-
 
 @ARTICLE{Li2011-ca,
   title    = "A statistical framework for {SNP} calling, mutation discovery,
@@ -39,7 +33,7 @@
   pages    = "2987--2993",
   month    =  nov,
   year     =  2011,
-  language = "en"
+  doi      = {10.1093/bioinformatics/btr509},
 }
 
 
@@ -56,7 +50,7 @@
   pages    = "2156--2158",
   month    =  aug,
   year     =  2011,
-  language = "en"
+  doi      = {10.1093/bioinformatics/btr330},
 }
 
 
@@ -72,7 +66,7 @@
   pages    = "68--74",
   month    =  oct,
   year     =  2015,
-  language = "en"
+  doi      = {10.1038/nature15393},
 }
 
 @ARTICLE{Speidel2019-ox,
@@ -86,7 +80,7 @@
   pages     = "1321--1329",
   month     =  sep,
   year      =  2019,
-  language  = "en"
+  doi       = {10.1038/s41588-019-0484-x},
 }
 
 @article{
@@ -99,8 +93,7 @@ number = {21},
 pages = {e2013798118},
 year = {2021},
 doi = {10.1073/pnas.2013798118},
[…]
