### 2023-02-10 af55ed5
maybe cyvcf2 wants zlib headers :grimacing:
