### 2023-02-11 f7e0146
idea: prioritize bioconda channel
