+++ .github/workflows/build-and-test.yml
@@ -13,11 +13,11 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest]
-        python-version: [3.7, 3.8, 3.9]
+        python-version: [3.7, 3.8, 3.9, 3.10, 3.11]
         include:
-          - os: linux
+          - os: ubuntu-latest
             compiler: gcc_linux-64
-          - os: macos
+          - os: macos-latest
             compiler: clang_osx-64
 
     defaults:
