+++ environment.yml
@@ -1,13 +1,0 @@
-name: mutyper
-channels:
-  - bioconda
-  - conda-forge
-  - defaults
-dependencies:
-  - python
-  - bcftools
-  - tabix
-  - pip
-  - pip:
-    - -r requirements.txt
-    - . # installs package from current directory
