+++ .github/workflows/docs-build-and-deploy.yml
@@ -30,8 +30,7 @@ jobs:
           make install
 
       - name: Build 🔧
-        # NOTE: must specify the shell so that conda init updates bashrc see:
-        #      https://github.com/conda-incubator/setup-miniconda#IMPORTANT
+        # must specify the shell so that conda init updates bashrc see: https://github.com/conda-incubator/setup-miniconda#IMPORTANT
         shell: bash -l {0}
         run: |
           make docs
