+++ .github/workflows/docs-build-and-deploy.yml
@@ -7,6 +7,12 @@ jobs:
   build-and-deploy:
     concurrency: ci-${{ github.ref }} # Recommended if you intend to make multiple deployments in quick succession.
     runs-on: ubuntu-latest
+
+    defaults:
+      # must specify the shell so that conda init updates bashrc see: https://github.com/conda-incubator/setup-miniconda#IMPORTANT
+      run:
+        shell: bash -el {0}
+
     steps:
       - name: Checkout 🛎️
         uses: actions/checkout@v2
@@ -30,8 +36,6 @@ jobs:
           make install
 
       - name: Build 🔧
-        # must specify the shell so that conda init updates bashrc see: https://github.com/conda-incubator/setup-miniconda#IMPORTANT
-        shell: bash -l {0}
         run: |
           make docs
 
