+++ docs/cite.rst
@@ -16,3 +16,6 @@ Citing ``mutyper``
      year     =  2020,
      language = "en"
     }
+
+.. todo::
+  replace with published version when available
