### 2023-02-12 d497a50
run workflow on the specified os (linux or mac)
