### 2023-02-12 a09fa9d
JOSS prereview (#40)
* JOSS status badge in README

* add missing DOIs to bib, and clean/simplify bib

* condensing some text to meet JOSS word limit

* now requires `python>=3.8`
