+++ paper/paper.md
@@ -40,38 +40,34 @@ affiliations:
    index: 4
  - name: The Roslin Institute, Royal (Dick) School of Veterinary Studies, University of Edinburgh, Easter Bush Campus, Midlothian, UK
    index: 5
-date: 24 January 2023
+date: 8 February 2023
 bibliography: paper.bib
 ---
 
 # Summary
 
-Genomics studies of the germline mutation process seek to elucidate the mutational forces that drive genetic variation and provide the raw material for adaptive evolution.
+The germline mutation process drives genetic variation and provides the raw material for adaptive evolution.
 Germline mutations arise from spontaneous DNA damage caused by environmental mutagens, or errors in DNA replication.
-Populations and species may experience distinct histories of mutational input due to differences in environmental exposure, life history (i.e. generation time), and heritable variation in the machinery controlling DNA replication fidelity.
+Populations and species may experience distinct mutational histories due to variation in environmental exposure, life history, and heritable variation in the machinery controlling DNA replication fidelity.
 
-Diverse mutational processes can reveal themselves in tell-tale mutation signatures left in the nucleotide sequence contexts in which they tend to be active.
+Mutational mechanisms often have _mutation signatures_ in terms of the nucleotide sequence contexts where they act.
 Population genomics has given increasing attention to nucleotide sequence context in the study of the germline mutation process (reviewed in @Carlson2020-hb).
 Single-nucleotide polymorphisms (SNPs) can be assigned to *mutation types* by the ancestral and derived nucleotide states and a window of local nucleotide context in the ancestral background.
 The *mutation spectrum* of an individual or population is the relative distribution of these mutation types.
 
 Inter- and intra-specific germline mutation spectrum variation has revealed a dynamic and evolving germline mutation process shaping modern genomic diversity.
-Parsing mutation spectra temporally (i.e. according to allele frequency) and spatially (i.e. in different genomic compartments) has revealed the history and present of various mutational processes, and applying such analysis to *de novo* mutation data may be clinically informative for rare or undiagnosed genetic diseases.
+Parsing mutation spectra temporally (via allele frequency) and spatially (via genomic annotations) has revealed the history and present of mutational processes, and applying such analysis to *de novo* mutation data may be clinically informative for rare or undiagnosed genetic diseases.
 
-Here we describe `mutyper`, a command-line utility and Python package that uses an ancestral genome estimate to assign mutation types to SNP data, compute mutation spectra for individuals and populations, and compute sample frequency spectra stratified by mutation type for population genetic inference.
+Here we describe `mutyper`, a command-line utility and Python package that assigns ancestrally polarized mutation types to SNP data, computes mutation spectra for individuals and populations, and computes sample frequency spectra stratified by mutation type for population genetic inference.
 Documentation is provided at [https://harrispopgen.github.io/mutyper](); source code is available at [https://github.com/harrispopgen/mutyper]().
 
 # Statement of need
 
-Despite many exciting findings to date, there is a lack of software for general-purpose germline mutation type partitioning and mutation spectrum generation from population-scale genomic variation data.
-There is a need for efficient and well-tested software for both larger bioinformatics pipelines and simple exploratory analysis.
-To address this need, we developed `mutyper`, an open-source command-line utility and Python package.
+Despite many exciting findings in this growing area, there is a lack of software for germline mutation type annotation and spectrum generation from population-scale genomic data.
+We developed `mutyper`, an open-source command-line utility and Python package, to address the field's need for efficient and well-tested software for both larger bioinformatics pipelines and exploratory analysis.
 
[…]
