+++ .github/workflows/build-and-test.yml
@@ -9,11 +9,22 @@ on:
 jobs:
   build:
     if: "!contains(github.event.commits[0].message, '[skip ci]')"
-    runs-on: ubuntu-latest
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest]
-        python-version: [3.7, 3.8, 3.9]
+        python-version: [3.8, 3.9, '3.10', 3.11]  # https://github.com/actions/runner/issues/1989
+        include:
+          - os: ubuntu-latest
+            compiler: gcc_linux-64
+          - os: macos-latest
+            compiler: clang_osx-64
+
+    runs-on: ${{ matrix.os }}
+
+    defaults:
+      # must specify the shell so that conda init updates bashrc see: https://github.com/conda-incubator/setup-miniconda#IMPORTANT
+      run:
+        shell: bash -el {0}
 
     steps:
       - name: Checkout 🛎️
@@ -25,27 +36,27 @@ jobs:
       - name: Install miniconda and create environment
         uses: conda-incubator/setup-miniconda@v2
         with:
+          miniconda-version: "latest"
           activate-environment: mutyper
-          environment-file: environment.yml
+          channels: bioconda,conda-forge,defaults
           python-version: ${{ matrix.python-version }}
           auto-activate-base: false
           auto-update-conda: true
 
+      - name: Install
+        run: |
+          conda install bcftools tabix pip ${{ matrix.compiler }}
+          make install
+
       - name: Lint
-        # NOTE: must specify the shell so that conda init updates bashrc see:
-        #      https://github.com/conda-incubator/setup-miniconda#IMPORTANT
-        shell: bash -l {0}
         run: make lint
 
       - name: Check format with black
-        shell: bash -l {0}
         run: black --check mutyper
 
       - name: Test
-        shell: bash -l {0}
         run: make test
 
       - name: Test docs build
-        shell: bash -l {0}
         run: |
           make docs
