+++ environment.yml
@@ -1,12 +1,0 @@
-name: mutyper
-channels:
-  - conda-forge
-  - defaults
-  - bioconda
-dependencies:
-  - bcftools
-  - tabix
-  - pip
-  - pip:
-    - -r requirements.txt
-    - . # installs package from current directory
