+++ paper/paper.bib
@@ -179,7 +179,7 @@ doi = {10.1073/pnas.2013798118},
   pages    = "3439--3444",
   month    =  mar,
   year     =  2015,
-  doi      = {10.1073/pnas.141865211},
+  doi      = {10.1073/pnas.1418652112},
 }
 
 @ARTICLE{Harris2017-fw,
