### 2023-02-16 #5102 @izabelcavassim
Hi Lorena,

I could revise the tool. What is the timeline for revision?

Thanks

On Wed, Feb 15, 2023 at 2:31 PM Lorena Pantano ***@***.***>
wrote:

> Hi @arq5x <https://github.com/arq5x>, @izabelcavassim
> <https://github.com/izabelcavassim>, would you like to review this tool?
> Let us know if you would have time! Here are the guidelines we use for
> reviewing:
> https://joss.readthedocs.io/en/latest/review_criteria.html#review-items.
> Thanks!
>
