### 2023-03-03 #5102 @WSDeWitt
Thanks @lpantano, and sorry about the difficulties. [These](https://github.com/openjournals/joss-reviews/issues/5102#issuecomment-1426925388) were my original reviewer suggestions. My criteria were:
- Work in computational population genomics
- Significant open-source collaborative software contributions
- I don't know them well personally or collaborate

However, I now realize I probably should have excluded AQ and BP at U Utah because one of the contributing authors recently moved there for a postdoc (sorry about that). If you've tried the other people on this list I can think of some more. Here are a few ideas:
- Kevin Thornton (UC Irvine): https://github.com/molpopgen
- Silas Tittes (UC Boulder): https://github.com/silastittes
- Ariella Gladstein (Embarkvet): https://github.com/agladstein
