### 2023-03-07 #5102 @lpantano
Hi @agladstein, authors mentioned you as possible reviewer, are you interested in reviewing this tool? The process looks like this https://joss.readthedocs.io/en/latest/review_criteria.html#review-items. Thanks!
