### 2023-03-07 #5227 @editorialbot
[REVIEW]: mutyper: assigning and summarizing mutation types for analyzing germline mutation spectra
**Submitting author:** @WSDeWitt (<a href="http://orcid.org/0000-0002-6802-9139">William DeWitt</a>)
**Repository:** https://github.com/harrispopgen/mutyper
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
