### 2023-03-07 #5227 @WSDeWitt
Just noting that the current software version is v7.0.1, as explained here: https://github.com/openjournals/joss-reviews/issues/5102#issuecomment-1426915639. Dependency issues arose in the originally submitted v7.0.0, but these are fixed in v7.0.1.
