### 2023-03-08 harrispopgen/mutyper#41 @vladsavelyev
Integration with Hail
I wonder if you happened to see [Hail Query](https://hail.is/tutorial.html), a framework that for processing variant data in scale. It was used e.g. to build the [gnomAD](https://gnomad.broadinstitute.org/) resource. It would be cool to support Hail matrix table in mutyper, or to contribute the mutyper methods to the Hail codebase itself (e.g. as part of https://hail.is/docs/0.2/methods/genetics.html#genetics)

Thanks for this amazing and well-implemented tool!
