### 2023-03-14 #5227 @lpantano
Hi @izabelcavassim, let me know if you need any help with the process. With this command `@editorialbot generate my checklist` you will have a guide on what steps to follow to complete the review.
