### 2023-03-14 #5227 @izabelcavassim
Ok I did get it generated, I will go through it this week!

Best,

Izabel
On Tue, Mar 14, 2023 at 7:49 AM Maria Izabel Cavassim Alves <
***@***.***> wrote:

> Thanks Lorena, let me try
> @editorialbot generate my checklist
>
> On Tue, Mar 14, 2023 at 7:43 AM Lorena Pantano ***@***.***>
> wrote:
>
>> Hi @izabelcavassim <https://github.com/izabelcavassim>, let me know if
>> you need any help with the process. With this command @editorialbot
>> generate my checklist you will have a guide on what steps to follow to
>> complete the review.
>>
>> —
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/5227#issuecomment-1468233105>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/AB3WMCO52HWOOHCEWWP6BS3W4B7ZZANCNFSM6AAAAAAVSVVTNI>
>> .
>> You are receiving this because you were mentioned.Message ID:
>> ***@***.***>
>>
> --
> Postdoctoral Researcher
> Dept. of Ecology and Evolutionary biology
> University of California - Los Angeles (UCLA) - United States
> Phone:  +1 (310) 593-1637
> E-mail: ***@***.***
> https://izabelcavassim.github.io/
> Linkedin profile
> <https://www.linkedin.com/profile/view?id=350503047&trk=nav_responsive_tab_profile>
>
>
--
Postdoctoral Researcher
Dept. of Ecology and Evolutionary biology
University of California - Los Angeles (UCLA) - United States
Phone:  +1 (310) 593-1637
E-mail: ***@***.***
https://izabelcavassim.github.io/
Linkedin profile
<https://www.linkedin.com/profile/view?id=350503047&trk=nav_responsive_tab_profile>
