### 2023-03-19 harrispopgen/mutyper#42 @izabelcavassim
Fix references in  https://github.com/harrispopgen/mutyper/blob/master/paper/paper.md
In the session [Applications](https://github.com/harrispopgen/mutyper/blob/master/paper/paper.md#applications) of their article, the authors give examples of where users have utilized mutyper but the authors have not shared the specific citations for these articles. I would like to see references for the work they have mentioned, specifically: @sasani2022 and @Vollger2022.
