### 2023-03-19 harrispopgen/mutyper#43 @izabelcavassim
Fix Documentation link
Please resolve this sentence in https://github.com/harrispopgen/mutyper/blob/master/paper/paper.md#summary:
"Documentation is provided at [https://harrispopgen.github.io/mutyper](https://github.com/harrispopgen/mutyper/blob/master/paper);"

to the actual link: https://harrispopgen.github.io/mutyper/
somehow is leading me to https://github.com/harrispopgen/mutyper/tree/master/paper
because of the wrong link
