### 2023-03-19 harrispopgen/mutyper#44 @izabelcavassim
Add Community Guidelines
I couldn't find anything related to community guidelines.

Can you add that?
