### 2023-03-19 harrispopgen/mutyper#45 @izabelcavassim
Add State of the field
State of the field: Do the authors describe how this software compares to other commonly-used packages?

I don't see anything describing the state of the field in here: https://github.com/harrispopgen/mutyper/blob/master/paper/paper.md
Can you please add that?

Maybe https://genomemedicine.biomedcentral.com/articles/10.1186/s13073-021-00854-6 could be one example of software comparable to mutyper.
