### 2023-03-19 harrispopgen/mutyper#46 @izabelcavassim
Example usage: reduce assumptions
In this specific session: https://harrispopgen.github.io/mutyper/notebooks/quickstart.html
the quickstart example assumes that the user is running a IPython (e.g., jupyter notebook), and assumes a relative path to the example data, but that does not need to be the case, please modify it so to reduce these assumptions.
