### 2023-03-19 harrispopgen/mutyper#42 @WSDeWitt
My understanding was that the markdown source file should have an empty References section, but then the JOSS editorialbot compiles the paper pdf using the bibtext file to fill in the references cited (with tags like `@sasani2022`). It looks like the references are rendering in the pdf compiled from the markdown + bibtext file, linked by the editorialbot comment here: https://github.com/openjournals/joss-reviews/issues/5227#issuecomment-1458443035.

Does this look ok?
