### 2023-03-20 harrispopgen/mutyper#42 @izabelcavassim
sorry that works, are you intending to leave the markdown in the repo too? Is there a way to print it out to markdown or only when rendered to pdf? I guess if you create references manually on markdown it would be overwritten by the pdf? I never tried so I don't know.
