### 2023-03-21 78ed669
add README to paper subdirectory, closes #42
