+++ paper/README.md
@@ -1,0 +1,5 @@
+# Paper
+
+This directory contains markdown source and bibtex references for a paper submitted to [JOSS](https://joss.theoj.org).
+A [GitHub Action](https://github.com/harrispopgen/mutyper/actions/workflows/draft-pdf.yml) triggered by push will compile a draft PDF and attach it as an artifact of the action run.
+The JOSS submission and open review process can be found [here](https://joss.theoj.org/papers/db2fb9b0e9fa1b380ff75655796e7bf1).
