### 2023-03-21 b4ecf4c
update docs quickstart, closes #46
