### 2023-04-02 a5d3fcd
contributor guidlines, closes #44
