+++ requirements.txt
@@ -6,8 +6,8 @@ flake8
 pytest
 sphinx
 sphinx-argparse
-sphinx_rtd_theme
+sphinx-rtd-theme
 sphinx-autodoc-typehints
 sphinxcontrib-programoutput
-myst_nb
+myst-nb
 sphinx-copybutton
