+++ paper/paper.bib
@@ -349,3 +349,80 @@ doi = {10.1073/pnas.2013798118},
   editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
   doi       = { 10.25080/Majora-92bf1922-00a }
 }
+
+@article{gehring2015somaticsignatures,
+  title={SomaticSignatures: inferring mutational signatures from single-nucleotide variants},
+  author={Gehring, Julian S and Fischer, Bernd and Lawrence, Michael and Huber, Wolfgang},
+  journal={Bioinformatics},
+  volume={31},
+  number={22},
+  pages={3673--3675},
+  year={2015},
+  publisher={Oxford University Press}
+}
+
+@article{rosenthal2016deconstructsigs,
+  title={DeconstructSigs: delineating mutational processes in single tumors distinguishes DNA repair deficiencies and patterns of carcinoma evolution},
+  author={Rosenthal, Rachel and McGranahan, Nicholas and Herrero, Javier and Taylor, Barry S and Swanton, Charles},
+  journal={Genome biology},
+  volume={17},
+  number={1},
+  pages={1--11},
+  year={2016},
+  publisher={BioMed Central}
+}
+
+@article{rosales2017signer,
+  title={signeR: an empirical Bayesian approach to mutational signature discovery},
+  author={Rosales, Rafael A and Drummond, Rodrigo D and Valieris, Renan and Dias-Neto, Emmanuel and Da Silva, Israel T},
+  journal={Bioinformatics},
+  volume={33},
+  number={1},
+  pages={8--16},
+  year={2017},
+  publisher={Oxford University Press}
+}
+
+@article{goncearenco2017exploring,
+  title={Exploring background mutational processes to decipher cancer genetic heterogeneity},
+  author={Goncearenco, Alexander and Rager, Stephanie L and Li, Minghui and Sang, Qing-Xiang and Rogozin, Igor B and Panchenko, Anna R},
+  journal={Nucleic acids research},
+  volume={45},
+  number={W1},
+  pages={W514--W522},
+  year={2017},
+  publisher={Oxford University Press}
+}
+
+@article{lee2018mutalisk,
+  title={Mutalisk: a web-based somatic MUTation AnaLyIS toolKit for genomic, transcriptional and epigenomic signatures},
+  author={Lee, Jongkeun and Lee, Andy Jinseok and Lee, June-Koo and Park, Jongkeun and Kwon, Youngoh and Park, Seongyeol and Chun, Hyonho and Ju, Young Seok and Hong, Dongwan},
+  journal={Nucleic acids research},
+  volume={46},
+  number={W1},
+  pages={W102--W108},
+  year={2018},
+  publisher={Oxford University Press}
+}
+
+@article{li2020using,
+  title={Using sigLASSO to optimize cancer mutation signatures jointly with sampling likelihood},
+  author={Li, Shantao and Crawford, Forrest W and Gerstein, Mark B},
+  journal={Nature communications},
+  volume={11},
+  number={1},
+  pages={3575},
+  year={2020},
+  publisher={Nature Publishing Group UK London}
+}
+
+@article{manders2022mutationalpatterns,
+  title={MutationalPatterns: the one stop shop for the analysis of mutational processes},
+  author={Manders, Freek and Brandsma, Arianne M and de Kanter, Jurrian and Verheul, Mark and Oka, Rurika and van Roosmalen, Markus J and van der Roest, Bastiaan and van Hoeck, Arne and Cuppen, Edwin and van Boxtel, Ruben},
+  journal={BMC genomics},
+  volume={23},
+  number={1},
+  pages={134},
+  year={2022},
+  publisher={Springer}
+}
