### 2023-04-03 harrispopgen/mutyper#42 @WSDeWitt
We haven't found a way to also render the references in the formatted markdown, but in PR #47 we've added a [README](https://github.com/harrispopgen/mutyper/tree/joss-review/paper)  in the `paper/` subdirectory to explain that the markdown paper and bibtex references are source files for a pdf paper build. We also provide a link to the JOSS review page (where the pdf paper would appear).
