### 2023-04-03 harrispopgen/mutyper#44 @WSDeWitt
In PR #47 we've added a [CONTRIBUTING](https://github.com/harrispopgen/mutyper/blob/joss-review/docs/CONTRIBUTING.md) page to the docs. After merging the PR, this page will also be rendered in the sphinx documentation alongside the developer documentation page.
