### 2023-04-03 harrispopgen/mutyper#45 @WSDeWitt
In PR #47 we've added several citations and expanded the second paragraph of the **Statement of need** section (the most recent pdf draft in that PR is downloadable as the artifact of [this](https://github.com/harrispopgen/mutyper/suites/11978100981/artifacts/628546822) action). Although there are several methods for extracting mutation spectra for somatic mutation data, they don't scale well to population genomic germline mutation data (with thousands of individuals and millions of variants), or are not configurable/flexible (e.g. to handle diploid or haploid data). Some of these issues are discussed in the Helmsman paper (Carlson et al.).
