### 2023-04-04 harrispopgen/mutyper#41 @WSDeWitt
Thanks for the positive assessment, and for pointing us to Hail—it looks like a very nice framework for computational genomics workflows.

Mutyper is currently built on top of cyvcf2 (for VCF processing) and pyfaidx (for FASTA processing), which work well for CLI streaming and random access with minimal memory requirements. These are both widely used packages, with hundreds of dependent repos or packages (see [here](https://github.com/brentp/cyvcf2/network/dependents) and [here](https://github.com/mdshw5/pyfaidx/network/dependents)).

We are open to including Hail interoperability on the roadmap for mutyper, but would likely need contributions from Hail experts for this, as well as an assessment of how this would improve, replace, or augment the current CLI and API based on cyvcf2 and pyfaidx.

Would you find it satisfactory if we leave this issue open with an "enhancement" tag for now?
