+++ paper/paper.md
@@ -106,7 +106,8 @@ As of this writing, `mutyper` is being used in several ongoing studies in multip
 
 # Acknowledgements
 
-Jedidiah Carlson and Sarah Hilton provided helpful comments.
+The authors thank reviewers Izabel Cavassim and Vlad Savelyev for comments and corrections.
+Jedidiah Carlson and Sarah Hilton provided comments on an early draft.
 WSD was supported by the National Institute Of Allergy And Infectious Diseases (F31AI150163), and a Fellowship in Understanding Dynamic and Multi-scale Systems from the James S. McDonnell Foundation.
 AT has been supported by the Institute Strategic Programme Grant BBS/E/D/10002070 from the Biotechnology and Biological Sciences Research Council (BBSRC).
 ACB was supported by the Biological Mechanisms of Healthy Aging Training Program, NIH T32AG066574.
