### 2023-04-06 8671cde
JOSS review (#47)
* add README to paper subdirectory, closes #42

* add missing urls, closes #43

* update docs quickstart, closes #46

* contributor guidlines, closes #44

* cite somatic mutation spectrum methods, closes #45

* acknowledge reviewers
