+++ docs/conf.py
@@ -29,18 +29,20 @@ extensions = [
     "sphinx.ext.mathjax",
     "sphinx.ext.napoleon",
     "sphinx.ext.viewcode",
-    "nbsphinx",
+    'sphinx_copybutton',
     "sphinx_autodoc_typehints",
     "sphinx.ext.todo",
     "sphinxarg.ext",
+    "myst_nb",
+    "sphinx.ext.doctest",
 ]
 
 todo_include_todos = True
 
 intersphinx_mapping = {
     'python': ('https://docs.python.org/3/', None),
-    'numpy': ('https://docs.scipy.org/doc/numpy/', None),
-    'scipy': ('https://docs.scipy.org/doc/scipy/reference/', None),
+    'numpy': ('https://numpy.org/doc/stable/', None),
+    'scipy': ('https://docs.scipy.org/doc/scipy/', None),
 }
 
 # Add any paths that contain templates here, relative to this directory.
@@ -57,7 +59,7 @@ master_doc = 'index'
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store', 'README.md']
 
 
 # The name of the Pygments (syntax highlighting) style to use.
@@ -67,78 +69,14 @@ pygments_style = None
 autosummary_generate = True
 napolean_use_rtype = False
 
-# -- Options for nbsphinx -----------------------------------------------------
 
-# Execute notebooks before conversion: 'always', 'never', 'auto' (default)
-# We execute all notebooks, exclude the slow ones using 'exclude_patterns'
-nbsphinx_execute = 'auto'
+# -- Options for myst ----------------------------------------------
+myst_heading_anchors = 3  # auto-generate 3 levels of heading anchors
+myst_enable_extensions = ['dollarmath']
+nb_execution_mode = "auto"  # NOTE: this is off because we are committing executed notebooks for testing
+nb_execution_allow_errors = False
+nb_merge_streams = True
 
-# Use this kernel instead of the one stored in the notebook metadata:
-# nbsphinx_kernel_name = 'python3'
-
-# List of arguments to be passed to the kernel that executes the notebooks:
-nbsphinx_execute_arguments = [
-    "--InlineBackend.figure_formats={'svg', 'pdf'}",
-    "--InlineBackend.rc={'figure.dpi': 96}",
-]
-
-# If True, the build process is continued even if an exception occurs:
-# nbsphinx_allow_errors = True
-
-
-# Controls when a cell will time out (defaults to 30; use -1 for no timeout):
-# nbsphinx_timeout = 180
-
-# Default Pygments lexer for syntax highlighting in code cells:
-# nbsphinx_codecell_lexer = 'ipython3'
-
-# Width of input/output prompts used in CSS:
-# nbsphinx_prompt_width = '8ex'
-
-# If window is narrower than this, input/output prompts are on separate lines:
-# nbsphinx_responsive_width = '700px'
-
-# This is processed by Jinja2 and inserted before each notebook
-nbsphinx_prolog = r"""
-{% set docname = 'docs/' + env.doc2path(env.docname, base=None) %}
-.. only:: html
-    .. role:: raw-html(raw)
-        :format: html
-    .. nbinfo::
-        Interactive online version:
[…]
