+++ docs/README.md
@@ -1,2 +1,4 @@
-For instructions on how to change and test notebooks, see
-[Update Documentation](https://harrispopgen.github.io/mutyper/developer.html#update-documentation). 
+Documentation
+==
+
+https://harrispopgen.github.io/mutyper
