+++ _sources/quickstart.rst.txt
@@ -1,0 +1,517 @@
+Quickstart
+==========
+
+This notebook demonstrates how to use ``mutyper`` for computing mutation
+type data for VCF/BCF data
+
+We will access some `example data <https://github.com/harrispopgen/mutyper/blob/master/docs/example_data/>`_ from a directory in the GitHub repo.
+
+Python API demo
+---------------
+
+.. testsetup::
+
+    # we move up to the root of the repo, so that example data paths in rendered docs
+    # can be given wrt the root
+    import os
+    os.chdir('..')
+
+.. testcode::
+
+    import mutyper
+
+Path to an ancestral FASTA file for human chromosome 1
+
+.. testcode::
+
+    fasta = 'docs/example_data/ancestor.fa'
+
+Print the first 3 lines
+
+.. testcode::
+
+    with open(fasta) as f:
+         for _ in range(3):
+             print(f.readline().rstrip())
+
+.. testoutput::
+
+    >1
+    TTTGTGGGAgACTATTCCTCCCATCTGCAACAGCTGCCCCTGCTGACTGCCCTTCTCTCCTCCCTCTCATCCCAGAGAAACAgGTCAGCTGGGAGCTTCT
+    GCCCCCACTGCCTAGGGACCAACAGGGGCAGGAGGCAGTCACTGACCCCGAGACGTTTGCATCCTGCACAGCTAGAGATCCTTTATTAAAAGCACACTGT
+
+The ``mutyper`` package has a single class ``Ancestor``. We instantiate
+an ``Ancestor`` object using our FASTA file - We’re interested in 3-mer
+context, which we can indicate with the ``k`` keyword argument (by
+default it will be 3 though).
+
+.. testcode::
+
+    ancestor = mutyper.Ancestor(fasta, k=3)
+
+We can inspect FASTA record names, showing that the FASTA contains two
+records, chromosome ``'1'`` and ``'2'``.
+
+.. testcode::
+
+    print(ancestor.keys())
+
+.. testoutput::
+
+    odict_keys(['1', '2'])
+
+
+and we can Pythonically slice sequences via fast random access without
+loading into memory (leaning on
+`pyfaidx <https://pythonhosted.org/pyfaidx/>`__ under the hood):
+
+.. testcode::
+
+    start = 100
+    end = 300
+    print(repr(ancestor['1'][start:end]))
+
+.. testoutput::
+
+    >1:101-300
+    GCCCCCACTGCCTAGGGACCAACAGGGGCAGGAGGCAGTCACTGACCCCGAGACGTTTGCATCCTGCACAGCTAGAGATCCTTTATTAAAAGCACACTGTTGGTTTCTGCTCAGTTCTTTATTGATTGGTGTGCCGTTTTCTCTGGAAGCCTCTTAAGAACACAGTGGCGCAGGCTGGGTGGAGCCGTCCCCCCATGgAG
+
+
+We can also access these ``FastaRecord`` slices as string-like biopython
+``Seq`` objects
+
+.. testcode::
+
+    print(ancestor['1'][start:end].seq)
+
+.. testoutput::
+
+    GCCCCCACTGCCTAGGGACCAACAGGGGCAGGAGGCAGTCACTGACCCCGAGACGTTTGCATCCTGCACAGCTAGAGATCCTTTATTAAAAGCACACTGTTGGTTTCTGCTCAGTTCTTTATTGATTGGTGTGCCGTTTTCTCTGGAAGCCTCTTAAGAACACAGTGGCGCAGGCTGGGTGGAGCCGTCCCCCCATGgAG
+
+
+
+The ``mutation_type`` method allows us to specify a SNP by the usual
+CHROM, POS, REF, ALT, and returns the correctly polarized mutation type
[…]
