### 2023-04-06 e728b35
add missing DOIs
