+++ paper/paper.bib
@@ -358,7 +358,8 @@ doi = {10.1073/pnas.2013798118},
   number={22},
   pages={3673--3675},
   year={2015},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/btv408}
 }
 
 @article{rosenthal2016deconstructsigs,
@@ -369,7 +370,8 @@ doi = {10.1073/pnas.2013798118},
   number={1},
   pages={1--11},
   year={2016},
-  publisher={BioMed Central}
+  publisher={BioMed Central},
+  doi={10.1186/s13059-016-0893-4},
 }
 
 @article{rosales2017signer,
@@ -380,7 +382,8 @@ doi = {10.1073/pnas.2013798118},
   number={1},
   pages={8--16},
   year={2017},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/btw572},
 }
 
 @article{goncearenco2017exploring,
@@ -391,7 +394,8 @@ doi = {10.1073/pnas.2013798118},
   number={W1},
   pages={W514--W522},
   year={2017},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/nar/gkx367},
 }
 
 @article{lee2018mutalisk,
@@ -402,7 +406,8 @@ doi = {10.1073/pnas.2013798118},
   number={W1},
   pages={W102--W108},
   year={2018},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/nar/gky406},
 }
 
 @article{li2020using,
@@ -413,7 +418,8 @@ doi = {10.1073/pnas.2013798118},
   number={1},
   pages={3575},
   year={2020},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41467-020-17388-x},
 }
 
 @article{manders2022mutationalpatterns,
@@ -424,5 +430,6 @@ doi = {10.1073/pnas.2013798118},
   number={1},
   pages={134},
   year={2022},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1186/s12864-022-08357-3},
 }
