+++ README.md
@@ -2,7 +2,7 @@
 
 A Python package and command line utility for annotating the local ancestral sequence context of biallelic SNPs
 
-[![build and test](https://github.com/harrispopgen/mutyper/actions/workflows/build-and-test.yml/badge.svg)](https://github.com/harrispopgen/mutyper/actions/workflows/build-and-test.yml)
+[![Build and test](https://github.com/harrispopgen/mutyper/actions/workflows/build-and-test.yml/badge.svg)](https://github.com/harrispopgen/mutyper/actions/workflows/build-and-test.yml)
 [![Docs build and deploy](https://github.com/harrispopgen/mutyper/actions/workflows/docs-build-and-deploy.yml/badge.svg)](https://github.com/harrispopgen/mutyper/actions/workflows/docs-build-and-deploy.yml)
 [![Python package](https://github.com/harrispopgen/mutyper/actions/workflows/python-publish.yml/badge.svg)](https://github.com/harrispopgen/mutyper/actions/workflows/python-publish.yml)
 [![status](https://joss.theoj.org/papers/db2fb9b0e9fa1b380ff75655796e7bf1/status.svg)](https://joss.theoj.org/papers/db2fb9b0e9fa1b380ff75655796e7bf1)
