### 2023-04-06 harrispopgen/mutyper#44 @WSDeWitt
Here is the new docs page: https://harrispopgen.github.io/mutyper/CONTRIBUTING.html
