### 2023-04-06 #5227 @WSDeWitt
Hi @lpantano,

We have now merged PR harrispopgen/mutyper#47 addressing reviewer-raised issues (listed below). The revised PDF is linked in the comment above. Upon merging the PR, many of the issues were automatically closed via commit messages. Please feel free to re-open issues that require further discussion or work to address. To see any additional comments summarizing our fix or response to each issue, you can click through to the issue pages:
* https://github.com/harrispopgen/mutyper/issues/41
* https://github.com/harrispopgen/mutyper/issues/42
* https://github.com/harrispopgen/mutyper/issues/43
* https://github.com/harrispopgen/mutyper/issues/44
* https://github.com/harrispopgen/mutyper/issues/45
* https://github.com/harrispopgen/mutyper/issues/46
* added a [line](https://github.com/harrispopgen/mutyper/blob/f092137e5c89b4fc951fe851d56cab0f0297cf58/paper/paper.md?plain=1#L109) to the paper acknowledging reviewers

Many thanks to reviewers @izabelcavassim and @vladsavelyev for scrutinizing our work.

  — mutyper authors
