### 2023-04-22 #5227 @WSDeWitt
@lpantano here is the info needed for the Post-Review Checklist:
- version: [v1.0.1](https://github.com/harrispopgen/mutyper/releases/tag/v1.0.1)
- DOI: https://doi.org/10.5281/zenodo.7855706
