+++ paper/paper.md
@@ -30,17 +30,17 @@ authors:
     orcid: 0000-0003-0302-2523
     affiliation: 3
 affiliations:
- - name: Department of Electrical Engineering & Computer Sciences, University of California, Berkeley, CA, USA
+ - name: Department of Electrical Engineering & Computer Sciences, University of California, Berkeley, CA, United States of America
    index: 1
- - name: Department of Bioengineering, University of Washington, Seattle, WA, USA
+ - name: Department of Bioengineering, University of Washington, Seattle, WA, United States of America
    index: 2
- - name: Department of Genome Sciences, University of Washington, Seattle, WA, USA
+ - name: Department of Genome Sciences, University of Washington, Seattle, WA, United States of America
    index: 3
- - name: Departments of Human Genetics and of Biomedical Informatics, University of Utah, Salt Lake City, UT, USA
+ - name: Departments of Human Genetics and of Biomedical Informatics, University of Utah, Salt Lake City, UT, United States of America
    index: 4
- - name: The Roslin Institute, Royal (Dick) School of Veterinary Studies, University of Edinburgh, Easter Bush Campus, Midlothian, UK
+ - name: The Roslin Institute, Royal (Dick) School of Veterinary Studies, University of Edinburgh, Easter Bush Campus, Midlothian, United Kingdom
    index: 5
-date: 8 February 2023
+date: 10 May 2023
 bibliography: paper.bib
 ---
 
