### 2023-05-10 #5227 @WSDeWitt
@Kevin-Mattheus-Moerman Thanks! I've made these changes and updated the GitHub version and Zenodo DOI as follows:
- GitHub [v1.0.2](https://github.com/harrispopgen/mutyper/releases/tag/v1.0.2)
- Zenodo DOI: [10.5281/zenodo.7921769](https://zenodo.org/record/7921769)
