### 2023-05-10 965a1a3
JOSS citation (#49)
* cff file

* add citation info to docs
