+++ _sources/cite.rst.txt
@@ -1,21 +1,25 @@
 Citing ``mutyper``
 ##################
 
-**Preprint:**
- W.S. DeWitt, *mutyper: assigning and summarizing mutation types for analyzing germline mutation spectra.* `bioRxiv, 2020.07.01.183392 <https://www.biorxiv.org/content/10.1101/2020.07.01.183392v1>`_ (2020).
+APA
+---
+
+DeWitt, W. S., Zhu, L., Vollger, M. R., Goldberg, M. E., Talenti, A., Beichman, A. C., & Harris, K. (2023). mutyper: assigning and summarizing mutation types for analyzing germline mutation spectra. *Journal of Open Source Software*, 8(85), 5227. https://doi.org/10.21105/joss.05227
+
+bibtex
+------
 
  .. code-block:: bibtex
 
-  @UNPUBLISHED{DeWitt2020-gp,
-     title    = "mutyper: assigning and summarizing mutation types for analyzing
-                 germline mutation spectra",
-     author   = "DeWitt, William S",
-     journal  = "bioRxiv",
-     pages    = "2020.07.01.183392",
-     month    =  jul,
-     year     =  2020,
-     language = "en"
-    }
-
-.. todo::
-  replace with published version when available
+  @article{DeWitt_mutyper_assigning_and_2023,
+  author = {DeWitt, William S. and Zhu, Luke and Vollger, Mitchell R. and Goldberg, Michael E. and Talenti, Andrea and Beichman, Annabel C. and Harris, Kelley},
+  doi = {10.21105/joss.05227},
+  journal = {Journal of Open Source Software},
+  month = may,
+  number = {85},
+  pages = {5227},
+  title = {{mutyper: assigning and summarizing mutation types for analyzing germline mutation spectra}},
+  url = {https://joss.theoj.org/papers/10.21105/joss.05227},
+  volume = {8},
+  year = {2023}
+  }
